move Fabric back to Trinkets and remove unnecessary and broken NF Accessories dep
