+++ Fabric/build.gradle
@@ -110,8 +110,6 @@ repositories {
     }
     // pehkui
     maven { url = "https://jitpack.io" }
-    // accessories
-    maven { url 'https://maven.wispforest.io/releases' }
 
     // If you have mod jar dependencies in ./libs, you can declare them as a repository like so:
     flatDir {
@@ -132,7 +130,7 @@ dependencies {
     // dear mod devs, don't put the platform in the middle of your version code
     modImplementation "vazkii.patchouli:Patchouli:1.21-$patchouliVersion-FABRIC-SNAPSHOT"
     modImplementation "com.samsthenerd.inline:inline-fabric:$minecraftVersion-$inlineVersion"
-    modImplementation "io.wispforest:accessories-fabric:$accessoriesVersion"
+    modImplementation "dev.emi:trinkets:${trinketsVersion}"
 
     modImplementation "org.ladysnake.cardinal-components-api:cardinal-components-base:$cardinalComponentsVersion"
 //    modImplementation "org.ladysnake.cardinal-components-api:cardinal-components-util:$cardinalComponentsVersion"
