+++ Fabric/gradle.properties
@@ -12,7 +12,7 @@ serializationHooksVersion=0.4.99999
 #entityReachVersion=2.3.0
 
 #gravityApiVersion=1.0.6
-trinketsVersion=3.7.0
+trinketsVersion=3.10.0
 
 modmenuVersion=7.0.1
