+++ Fabric/src/main/java/at/petrak/hexcasting/fabric/interop/trinkets/LensAccessoryRenderer.java
@@ -1,10 +1,10 @@
-package at.petrak.hexcasting.fabric.interop.accessories;
+package at.petrak.hexcasting.fabric.interop.trinkets;
 
 import at.petrak.hexcasting.common.lib.HexItems;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.math.Axis;
-import io.wispforest.accessories.api.client.AccessoryRenderer;
-import io.wispforest.accessories.api.slot.SlotReference;
+import dev.emi.trinkets.api.SlotReference;
+import dev.emi.trinkets.api.client.TrinketRenderer;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.model.EntityModel;
 import net.minecraft.client.model.PlayerModel;
@@ -19,12 +19,11 @@ import net.minecraft.world.item.ItemStack;
  * @author WireSegal
  * Created at 9:50 AM on 7/25/22.
  */
-public class LensAccessoryRenderer implements AccessoryRenderer {
+public class LensAccessoryRenderer implements TrinketRenderer {
     @Override
-    @SuppressWarnings("unchecked")
-    public <M extends LivingEntity> void render(ItemStack stack, SlotReference slotReference, PoseStack matrices, EntityModel<M> model, MultiBufferSource multiBufferSource, int light, float v, float v1, float v2, float v3, float v4, float v5) {
+    public void render(ItemStack stack, SlotReference slotReference, EntityModel<? extends LivingEntity> model, PoseStack matrices, MultiBufferSource multiBufferSource, int light, LivingEntity entity, float limbAngle, float limbDistance, float tickDelta, float animationProgress, float headYaw, float headPitch) {
         if (stack.is(HexItems.SCRYING_LENS) &&
-                model instanceof PlayerModel playerModel && slotReference.entity() instanceof Player) {
+                model instanceof PlayerModel playerModel && entity instanceof Player) {
 
             // from https://github.com/Creators-of-Create/Create/blob/ee33823ed0b5084af10ed131a1626ce71db4c07e/src/main/java/com/simibubi/create/compat/curios/GogglesCurioRenderer.java
