+++ Fabric/src/main/resources/data/trinkets/tags/item/head/face.json
