+++ gradle.properties
@@ -18,7 +18,6 @@ fabric_loader_version=0.16.14
 # haha not anymore it's from mavenLocal
 paucalVersion=0.7.1-noelle
 patchouliVersion=88
-accessoriesVersion=1.1.0-beta.16+1.21.1
 
 emiVersion=1.1.18+1.21.1
